[주문] 주문생성 - 2

+++ src/main/java/com/plateer/ec1/promotion/vo/Product.java
@@ -13,7 +13,7 @@ public class Product {
     @NotEmpty
     private String productNo; // 상품번호
     private long productAmt; // 가격
-    private int productCnt; // 수량
+    private Long productCnt; // 수량
     private String productItemNo; // 단품번호
     private Long prmNo;
     private Long cpnIssNo;
